Forgot password workflow

+++ src/server/src/auth/services/auth.service.ts
@@ -11,7 +11,7 @@ import { IUser, JWT, UserId } from "../../../../shared/entities";
 import { EMAIL_VERIFICATION_TOKEN_LENGTH, FROM_EMAIL } from "../../common/constants";
 import { User } from "../../users/entities/user.entity";
 import { UsersService } from "../../users/services/users.service";
-import { EmailVerification } from "../entities/email-verification.entity";
+import { EmailVerification, VerificationType } from "../entities/email-verification.entity";
 
 function generateEmailToken(): Promise<string> {
   return new Promise((resolve, reject) => {
@@ -24,6 +24,8 @@ function generateEmailToken(): Promise<string> {
     });
   });
 }
+
+const EMAIL_EXPIRATION_IN_MS = 60 * 60 * 24 * 1000;
 
 @Injectable()
 export class AuthService {
@@ -67,22 +69,8 @@ export class AuthService {
     return this.jwtService.sign(payload);
   }
 
-  async sendVerificationEmail(user: User): Promise<void> {
-    const emailToken = await generateEmailToken();
-    const data = {
-      email: user.email,
-      emailToken,
-      timestamp: new Date()
-    };
-
-    await this.emailVerificationRepo
-      .createQueryBuilder()
-      .insert()
-      .into(EmailVerification)
-      .values(data)
-      .onConflict(`("email") DO UPDATE SET "emailToken" = :emailToken, "timestamp" = :timestamp`)
-      .setParameters(data)
-      .execute();
+  async sendInitialVerificationEmail(user: User): Promise<void> {
+    const emailToken = this.sendEmailVerification(user, VerificationType.INITIAL);
 
     const info = await this.mailerService.sendMail({
       to: user.email,
@@ -100,24 +88,87 @@ export class AuthService {
     return;
   }
 
-  async verifyEmail(token: string): Promise<User | undefined> {
-    const emailVerif = await this.emailVerificationRepo.findOne({
-      emailToken: token
+  async sendPasswordResetEmail(user: User): Promise<void> {
+    const emailToken = await this.sendEmailVerification(user, VerificationType.FORGOT_PASSWORD);
+
+    const info = await this.mailerService.sendMail({
+      to: user.email,
+      from: FROM_EMAIL,
+      subject: "Reset your DistrictBuilder password",
+      context: {
+        user,
+        url: `${process.env.CLIENT_URL}/password-reset/${emailToken}`
+      },
+      template: "passwordreset"
     });
-    if (emailVerif) {
+
+    this.logger.debug(info.envelope, info.message.toString());
+
+    return;
+  }
+
+  async verifyInitialEmail(token: string): Promise<User | undefined> {
+    return this.verifyEmail(token, VerificationType.INITIAL, user => {
+      /* tslint:disable:no-object-mutation */
+      user.isEmailVerified = true;
+      /* tslint:enable */
+    });
+  }
+
+  async resetPassword(token: string, password: string): Promise<User | undefined> {
+    return this.verifyEmail(token, VerificationType.FORGOT_PASSWORD, user => {
+      /* tslint:disable:no-object-mutation */
+      // Token-based password reset goes through email and so will verify email
+      // address if that hasn't happened yet
+      user.isEmailVerified = true;
+      return user.setPassword(password);
+      /* tslint:enable */
+    });
+  }
+
+  private async sendEmailVerification(user: User, type: VerificationType): Promise<string> {
+    const emailToken = await generateEmailToken();
+    const data = {
+      email: user.email,
+      emailToken,
+      timestamp: new Date(),
+      type
+    };
+
+    await this.emailVerificationRepo
+      .createQueryBuilder()
+      .insert()
+      .into(EmailVerification)
+      .values(data)
+      .onConflict(
+        `("email", "type") DO UPDATE SET "emailToken" = :emailToken, "timestamp" = :timestamp`
+      )
+      .setParameters(data)
+      .execute();
+
+    return emailToken;
+  }
+
+  private async verifyEmail(
+    emailToken: string,
+    type: VerificationType,
+    updateUser: (user: User) => void | Promise<void>
+  ): Promise<User | undefined> {
+    const emailVerif = await this.emailVerificationRepo.findOne({ emailToken, type });
+    const now = new Date();
+    const lastValidTimestamp = new Date(now.getTime() - EMAIL_EXPIRATION_IN_MS);
+    if (emailVerif && emailVerif.timestamp > lastValidTimestamp) {
       const userFromDb = await this.usersService.findOne({
         email: emailVerif.email
       });
       if (userFromDb) {
-        /* tslint:disable:no-object-mutation */
-        userFromDb.isEmailVerified = true;
+        await updateUser(userFromDb);
         let savedUser;
         await getManager().transaction(async transactionalEntityManager => {
           savedUser = await transactionalEntityManager.save(userFromDb);
           await transactionalEntityManager.remove(emailVerif);
         });
         return savedUser;
-        /* tslint:enable */
       }
     }
   }
